requirements.txt
pytest
pytest-html
allure-pytest
pytest-order
requests
python-dotenv
pytest-xdist